refactor(editor): consolidate $$ delimiter logic into shared module
Extract MATH_DELIMITERS, INLINE_MATH_REGEX, and isEscaped() into
math-delimiters.ts so the markdown-it parser and tiptap input/paste
rules share identical escaping semantics. The regex now also rejects
empty and whitespace-only content via a negative lookahead.
